Align filenames with Kokkos naming conventions and change references t#
