+++ src/resilience/exec_space/openMP/CMakeLists.txt
@@ -1,4 +1,4 @@
 target_sources(resilience PRIVATE
-               ${CMAKE_CURRENT_LIST_DIR}/ResOpenMP.cpp
-               ${CMAKE_CURRENT_LIST_DIR}/OpenMPResSubscriber.cpp
+               ${CMAKE_CURRENT_LIST_DIR}/Resilient_OpenMP.cpp
+               ${CMAKE_CURRENT_LIST_DIR}/Resilient_OpenMP_Subscriber.cpp
                )
